Refactor shares field to use BigInteger in Delegator and Validator classes for improved precision

+++ src/main/java/com/github/pwrlabs/pwrj/entities/Delegator.java
@@ -3,11 +3,13 @@ package com.github.pwrlabs.pwrj.entities;
 import lombok.Getter;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigInteger;
+
 @Getter
 @SuperBuilder
 public class Delegator {
     private final String address;
     private final String validatorAddress;
-    private final long shares;
+    private final BigInteger shares;
     private final long delegatedPWR;
 }
